Decompose review utility helpers
