+++ src/issue-131/evidence-matrix.ts
@@ -98,8 +98,8 @@ export type Issue131CheckId = typeof ISSUE_131_CHECK_IDS[number];
 
 export const ISSUE_131_SOURCE_PATHS = [
   "src/handlers/review.ts",
-  "src/review-plan/review-plan.ts",
-  "src/lib/review-utils.ts",
+  "src/review-orchestration/review-plan.ts",
+  "src/lib/review-details-formatting.ts",
   "src/execution/config.ts",
   "src/review-graph/validation.ts",
   "src/review-graph/graph-validation-status.ts",
@@ -271,8 +271,8 @@ export function validateIssue131EvidencePath(path: string): { valid: true } | { 
 export function evaluateIssue131EvidenceMatrix(readers: Issue131EvidenceReaders): Issue131EvidenceMatrixReport {
   const source = {
     review: readers.readFileText("src/handlers/review.ts") ?? "",
-    reviewPlan: readers.readFileText("src/review-plan/review-plan.ts") ?? "",
-    reviewUtils: readers.readFileText("src/lib/review-utils.ts") ?? "",
+    reviewPlan: readers.readFileText("src/review-orchestration/review-plan.ts") ?? "",
+    reviewDetailsFormatting: readers.readFileText("src/lib/review-details-formatting.ts") ?? "",
     config: readers.readFileText("src/execution/config.ts") ?? "",
     validation: readers.readFileText("src/review-graph/validation.ts") ?? "",
     graphValidationStatus: readers.readFileText("src/review-graph/graph-validation-status.ts") ?? "",
@@ -303,7 +303,7 @@ export function evaluateIssue131EvidenceMatrix(readers: Issue131EvidenceReaders)
   };
 }
 
-function classifyRow(definition: RowDefinition, source: { review: string; reviewPlan: string; reviewUtils: string; config: string; validation: string; graphValidationStatus: string; packageJson: string }, handoffRows: readonly Issue131DeferredHandoffRow[]): Issue131MatrixRow {
+function classifyRow(definition: RowDefinition, source: { review: string; reviewPlan: string; reviewDetailsFormatting: string; config: string; validation: string; graphValidationStatus: string; packageJson: string }, handoffRows: readonly Issue131DeferredHandoffRow[]): Issue131MatrixRow {
   if (definition.deferredTo) {
     const handoff = handoffRows.find((row) => row.rowId === definition.id);
     const deferredTo = handoff ? {
@@ -318,23 +318,22 @@ function classifyRow(definition: RowDefinition, source: { review: string; review
     case "review-plan-contract": {
       const hasTypedContract = /export\s+type\s+ReviewPlan\b|export\s+interface\s+ReviewPlan\b/.test(source.reviewPlan);
       const hasBuilder = /export\s+function\s+buildReviewPlan\b/.test(source.reviewPlan);
-      const hasStableHash = /stableHash\s*:\s*string/.test(source.reviewPlan) && /REVIEW_PLAN_HASH_PREFIX|review-plan:v/.test(source.reviewPlan);
-      const hasSafeDiagnosticProjection = /export\s+function\s+summarizeReviewPlanForDiagnostics\b/.test(source.reviewPlan)
-        && /planHash/.test(source.reviewPlan)
-        && /assertNoForbiddenRawFields/.test(source.reviewPlan);
+      const hasStableHash = /hash\s*:\s*string/.test(source.reviewPlan) && /hashCanonical|createHash/.test(source.reviewPlan);
+      const hasSafeDiagnosticProjection = /export\s+function\s+toReviewPlanDetailsSummary\b/.test(source.reviewPlan)
+        && /Review plan: ready/.test(source.reviewPlan)
+        && /sanitizeSummaryToken/.test(source.reviewPlan);
       return hasTypedContract && hasBuilder && hasStableHash && hasSafeDiagnosticProjection
-        ? makeRow(definition, "complete", [{ path: "src/review-plan/review-plan.ts", reason: "Exports the typed ReviewPlan contract, stable hash, builder, and safe diagnostic projection." }])
-        : makeRow(definition, "missing", [], ["No exported typed ReviewPlan contract with stable hash, builder, and safe diagnostic projection was found in src/review-plan/review-plan.ts."]);
+        ? makeRow(definition, "complete", [{ path: "src/review-orchestration/review-plan.ts", reason: "Exports the typed ReviewPlan contract, stable hash, builder, and safe Review Details projection." }])
+        : makeRow(definition, "missing", [], ["No exported typed ReviewPlan contract with stable hash, builder, and safe Review Details projection was found in src/review-orchestration/review-plan.ts."]);
     }
     case "normal-handler-plan-construction": {
-      const hasReviewPlanImport = /from\s+[\"']\.\.\/review-plan\/review-plan\.ts[\"']/.test(source.review)
+      const hasReviewPlanImport = /from\s+[\"']\.\.\/review-orchestration\/review-plan\.ts[\"']/.test(source.review)
         && /\bbuildReviewPlan\b/.test(source.review)
-        && /\bsummarizeReviewPlanForDiagnostics\b/.test(source.review);
+        && /\btoReviewPlanDetailsSummary\b/.test(source.review);
       const construction = findReviewPlanConstruction(source.review);
       const constructsBeforePublication = construction.found && construction.beforePublication;
-      const hasSafeStableHashProjection = /reviewPlanSummarizer\s*\(\s*reviewPlan/.test(source.review)
-        || /\bplanHash\b/.test(source.review)
-        || /\bstableHash\b/.test(source.review);
+      const hasSafeStableHashProjection = /toReviewPlanDetailsSummary\s*\(\s*reviewPlan\s*\)/.test(source.review)
+        || /\breviewPlan\.hash\b/.test(source.review);
       if (hasReviewPlanImport && constructsBeforePublication && hasSafeStableHashProjection) {
         return makeRow(definition, "complete", [{ path: "src/handlers/review.ts", reason: "Normal review flow imports, constructs, and logs a safe ReviewPlan projection before nearby publication side effects." }]);
       }
@@ -345,22 +344,22 @@ function classifyRow(definition: RowDefinition, source: { review: string; review
     }
     case "review-details-plan-summary": {
       const projection = findReviewDetailsPlanProjection(source.reviewPlan);
-      const formatter = findReviewDetailsPlanFormatter(source.reviewUtils);
+      const formatter = findReviewDetailsPlanFormatter(source.reviewDetailsFormatting);
       const handler = findReviewDetailsPlanHandlerWiring(source.review);
       const hasReviewDetails = source.review.includes("formatReviewDetailsSummary")
-        || source.reviewUtils.includes("<summary>Review Details</summary>");
+        || source.reviewDetailsFormatting.includes("<summary>Review Details</summary>");
 
       if (projection.complete && formatter.complete && handler.complete) {
         return makeRow(definition, "complete", [
-          { path: "src/review-plan/review-plan.ts", reason: "Exports the compact public ReviewPlan Review Details projection with hash, route, scope, context, gate, budget, and publish-policy fields." },
-          { path: "src/lib/review-utils.ts", reason: "Formats the public ReviewPlan projection into a bounded Review Details line without raw review artifacts." },
+          { path: "src/review-orchestration/review-plan.ts", reason: "Exports the compact public ReviewPlan Review Details projection with hash, route, budget, gate, policy, graph, candidate, and doctrine fields." },
+          { path: "src/lib/review-details-formatting.ts", reason: "Formats the public ReviewPlan projection into a bounded Review Details line without raw review artifacts." },
           { path: "src/handlers/review.ts", reason: "Normal review publication passes the ReviewPlan Review Details projection to formatReviewDetailsSummary with fail-open warning logs." },
         ]);
       }
 
       const partialEvidence: Issue131Evidence[] = [];
-      if (projection.present) partialEvidence.push({ path: "src/review-plan/review-plan.ts", reason: "Public ReviewPlan Review Details projection exists but is not fully proven." });
-      if (formatter.present) partialEvidence.push({ path: "src/lib/review-utils.ts", reason: "Review Details formatter mentions ReviewPlan summary but bounded safe output is not fully proven." });
+      if (projection.present) partialEvidence.push({ path: "src/review-orchestration/review-plan.ts", reason: "Public ReviewPlan Review Details projection exists but is not fully proven." });
+      if (formatter.present) partialEvidence.push({ path: "src/lib/review-details-formatting.ts", reason: "Review Details formatter mentions ReviewPlan summary but bounded safe output is not fully proven." });
       if (handler.present || hasReviewDetails) partialEvidence.push({ path: "src/handlers/review.ts", reason: "Review Details publication exists or mentions ReviewPlan summary, but source-owned projection-to-formatter wiring is incomplete." });
 
       const failureReasons = [...projection.reasons, ...formatter.reasons, ...handler.reasons];
@@ -397,7 +396,7 @@ function classifyRow(definition: RowDefinition, source: { review: string; review
         return makeRow(definition, "complete", [
           { path: "src/review-graph/graph-validation-status.ts", reason: "Shared status mapper owns bounded graph-validation pre/runtime states, reasons, counts, and fail-open failure status." },
           { path: "src/handlers/review.ts", reason: "Normal review path feeds graph-validation into ReviewPlan gates and bounded skipped, unavailable, applied, and failure runtime logs." },
-          { path: "src/review-plan/review-plan.ts", reason: "ReviewPlan gate taxonomy supports enabled, applied, skipped, and unavailable graph-validation gate states." },
+          { path: "src/review-orchestration/review-plan.ts", reason: "ReviewPlan gate taxonomy supports enabled, applied, skipped, and unavailable graph-validation gate states." },
         ]);
       }
       if (statusProbe.present) {
@@ -501,41 +500,40 @@ function findTruthfulGraphValidationStatus(reviewSource: string, reviewPlanSourc
   const evidence: Issue131Evidence[] = [];
   if (/GraphValidationRuntimeStatus|GRAPH_VALIDATION_GATE|graphValidationAppliedRuntimeStatus/.test(status)) evidence.push({ path: "src/review-graph/graph-validation-status.ts", reason: "Graph-validation status mapper exists but the full evidence contract is not proven." });
   if (/graphValidationGateForReviewPlan|graphValidationSkippedRuntimeStatus|graphValidationAppliedRuntimeStatus|graphValidationThrownRuntimeStatus/.test(review)) evidence.push({ path: "src/handlers/review.ts", reason: "Review handler references graph-validation status helpers but complete ReviewPlan/runtime surfacing is not proven." });
-  if (/GATE_STATUSES|ReviewPlanGateStatus/.test(reviewPlan)) evidence.push({ path: "src/review-plan/review-plan.ts", reason: "ReviewPlan gate status taxonomy exists but graph-validation status wiring is incomplete." });
+  if (/GraphValidationPlanStatus|graphValidation/.test(reviewPlan)) evidence.push({ path: "src/review-orchestration/review-plan.ts", reason: "ReviewPlan graph-validation status taxonomy exists but graph-validation status wiring is incomplete." });
   return { ...probe, evidence };
 }
 
 function findReviewDetailsPlanProjection(reviewPlanSource: string): SourceEvidenceProbe {
-  const present = /summarizeReviewPlanForReviewDetails|ReviewPlanReviewDetailsSummary|review-plan-review-details/.test(reviewPlanSource);
+  const present = /toReviewPlanDetailsSummary|ReviewPlanDetailsSummary|Review plan: ready/.test(reviewPlanSource);
   return makeProbe(present, [
-    [/export\s+type\s+ReviewPlanReviewDetailsSummary\b/.test(reviewPlanSource), "No exported ReviewPlanReviewDetailsSummary type was found in source."],
-    [/export\s+function\s+summarizeReviewPlanForReviewDetails\s*\(\s*plan\s*:\s*ReviewPlan/.test(reviewPlanSource), "No source-owned ReviewPlan-to-Review Details summary projection was found."],
-    [/gate\s*:\s*["']review-plan-review-details["']/.test(reviewPlanSource), "Review Details projection does not identify the public review-plan gate."],
-    [["planHash", "route", "scope", "contextSources", "gates", "budgets", "publishPolicy"].every((field) => reviewPlanSource.includes(field)), "Review Details projection does not include hash, route, scope, context, gates, budgets, and publish policy."],
-    [/MAX_PUBLIC_REVIEW_DETAILS|omitted(?:Path|Source|Gate)Count|sanitizePublicReviewDetailsString/.test(reviewPlanSource), "Review Details projection does not prove public bounds, omitted counts, and string sanitization."],
+    [/export\s+type\s+ReviewPlanDetailsSummary\b/.test(reviewPlanSource), "No exported ReviewPlanDetailsSummary type was found in source."],
+    [/export\s+function\s+toReviewPlanDetailsSummary\s*\(\s*plan\s*:\s*ReviewPlan\s*\|\s*DegradedReviewPlan/.test(reviewPlanSource), "No source-owned ReviewPlan-to-Review Details line projection was found."],
+    [/Review plan: ready/.test(reviewPlanSource) && /Review plan: degraded/.test(reviewPlanSource), "Review Details projection does not render ready and degraded Review plan lines."],
+    [["hash", "route", "task", "files", "lines", "budget", "gates", "publish", "graph", "candidates", "doctrine"].every((field) => reviewPlanSource.includes(field)), "Review Details projection does not include hash, route, task, files, lines, budget, gates, publish, graph, candidates, and doctrine truth."],
+    [/boundSummary|sanitizeSummaryToken/.test(reviewPlanSource), "Review Details projection does not prove bounded/sanitized visible output."],
   ]);
 }
 
-function findReviewDetailsPlanFormatter(reviewUtilsSource: string): SourceEvidenceProbe {
-  const present = /ReviewPlanReviewDetailsFormatterSummary|formatReviewPlanReviewDetailsLine|reviewPlanSummary/.test(reviewUtilsSource);
-  const formatterBody = extractFunctionBody(reviewUtilsSource, "formatReviewPlanReviewDetailsLine");
+function findReviewDetailsPlanFormatter(reviewDetailsFormatterSource: string): SourceEvidenceProbe {
+  const present = /formatReviewPlanDetailsLine|reviewPlan\?: ReviewPlanDetailsSummary/.test(reviewDetailsFormatterSource);
+  const formatterBody = extractFunctionBody(reviewDetailsFormatterSource, "formatReviewPlanDetailsLine");
   const forbiddenVisibleTokens = ["rawPrompt", "modelPrompt", "rawModelOutput", "rawDiff", "candidatePayload", "candidateFindingPayload", "commentBody", "rawCommentBody", "apiKey", "password", "secret"];
   return makeProbe(present, [
-    [/export\s+type\s+ReviewPlanReviewDetailsFormatterSummary\b/.test(reviewUtilsSource), "No formatter-facing public ReviewPlan summary type was found."],
-    [formatterBody !== "" && formatterBody.includes("- Review Plan: hash="), "Review Details formatter does not render a compact Review Plan line."],
-    [["route=", "scope=", "contexts=", "gates=", "budget=", "publish="].every((token) => formatterBody.includes(token)), "Review Plan formatter line does not include route, scope, context, gate, budget, and publish-policy truth."],
-    [/REVIEW_PLAN_DETAILS_MAX_|sanitizeReviewPlanDetailsValue|formatReviewPlanDetailsRepresentativeList/.test(reviewUtilsSource), "Review Plan formatter does not prove bounded/sanitized visible output."],
+    [/ReviewPlanDetailsSummary/.test(reviewDetailsFormatterSource), "Formatter does not consume the canonical ReviewPlanDetailsSummary type."],
+    [formatterBody !== "" && /reviewPlan\?\.text/.test(formatterBody), "Review Details formatter does not render the source-owned Review plan line."],
+    [!/(ReviewPlanReviewDetailsFormatterSummary|formatReviewPlanReviewDetailsLine|reviewPlanSummary)/.test(reviewDetailsFormatterSource), "Review Details formatter still contains the legacy reviewPlanSummary path."],
     [!forbiddenVisibleTokens.some((token) => formatterBody.includes(token)), "Review Plan formatter visible output contains raw review artifact field names or canaries."],
   ]);
 }
 
 function findReviewDetailsPlanHandlerWiring(reviewSource: string): SourceEvidenceProbe {
-  const present = /summarizeReviewPlanForReviewDetails|reviewPlanReviewDetailsSummarizer|reviewPlanSummary/.test(reviewSource);
+  const present = /toReviewPlanDetailsSummary|reviewPlanDetailsSummary|reviewPlan\s*:/.test(reviewSource);
   return makeProbe(present, [
-    [/summarizeReviewPlanForReviewDetails/.test(reviewSource), "Review handler does not import/use the public ReviewPlan Review Details projection."],
-    [/reviewPlanReviewDetailsSummarizer\s*\(\s*reviewPlan\s*\)/.test(reviewSource), "Review handler does not derive the public ReviewPlan summary from the typed ReviewPlan instance."],
-    [/reviewPlanSummary\s*:\s*buildReviewPlanReviewDetailsSummary\s*\(\s*\)/.test(reviewSource), "Review handler does not pass the public ReviewPlan summary into formatReviewDetailsSummary."],
-    [/review-details-projection-failed/.test(reviewSource) && /logger\.warn/.test(reviewSource) && /fail-open/i.test(reviewSource), "Review handler does not prove bounded fail-open warning behavior for projection failures."],
+    [/toReviewPlanDetailsSummary/.test(reviewSource), "Review handler does not import/use the public ReviewPlan Review Details projection."],
+    [/toReviewPlanDetailsSummary\s*\(\s*reviewPlan\s*\)/.test(reviewSource), "Review handler does not derive the public ReviewPlan summary from the typed ReviewPlan instance."],
+    [/reviewPlan\s*:\s*reviewPlanDetailsSummary/.test(reviewSource), "Review handler does not pass the public ReviewPlan summary into formatReviewDetailsSummary."],
+    [/createDegradedReviewPlan/.test(reviewSource) && /builder-error/.test(reviewSource), "Review handler does not prove bounded fail-open ReviewPlan degradation behavior."],
   ]);
 }
