+++ src/lib/review-finding-metadata.ts
@@ -1,0 +1,120 @@
+export type ReviewArea = "security" | "correctness" | "performance" | "style" | "documentation";
+export type FindingSeverity = "critical" | "major" | "medium" | "minor";
+export type FindingCategory = "security" | "correctness" | "performance" | "style" | "documentation";
+export type ConfidenceBand = "high" | "medium" | "low";
+
+export function toConfidenceBand(confidence: number): ConfidenceBand {
+  if (confidence >= 75) return "high";
+  if (confidence >= 50) return "medium";
+  return "low";
+}
+
+export function fingerprintFindingTitle(title: string): string {
+  const normalized = title
+    .trim()
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, " ")
+    .replace(/\s+/g, " ");
+
+  let hash = 2166136261;
+  for (let i = 0; i < normalized.length; i += 1) {
+    hash ^= normalized.charCodeAt(i);
+    hash = Math.imul(hash, 16777619);
+  }
+
+  const unsigned = hash >>> 0;
+  return `fp-${unsigned.toString(16).padStart(8, "0")}`;
+}
+
+export function parseSeverityCountsFromBody(body: string): {
+  critical: number;
+  major: number;
+  medium: number;
+  minor: number;
+} {
+  const countMatches = (tag: string) => {
+    const regex = new RegExp(`\\[${tag}\\]`, "gi");
+    return (body.match(regex) || []).length;
+  };
+  return {
+    critical: countMatches("CRITICAL"),
+    major: countMatches("MAJOR"),
+    medium: countMatches("MEDIUM"),
+    minor: countMatches("MINOR"),
+  };
+}
+
+export function normalizeSeverity(value: string | undefined): FindingSeverity | null {
+  if (!value) return null;
+  const normalized = value.trim().toLowerCase();
+  if (normalized === "critical" || normalized === "major" || normalized === "medium" || normalized === "minor") {
+    return normalized;
+  }
+  return null;
+}
+
+export function normalizeCategory(value: string | undefined): FindingCategory {
+  if (!value) return "correctness";
+  const normalized = value.trim().toLowerCase();
+  if (normalized === "security") return "security";
+  if (normalized === "correctness" || normalized === "error-handling") return "correctness";
+  if (normalized === "performance" || normalized === "resource-management" || normalized === "concurrency") {
+    return "performance";
+  }
+  if (normalized === "style") return "style";
+  if (normalized === "documentation") return "documentation";
+  return "correctness";
+}
+
+export function parseInlineCommentMetadata(body: string): {
+  severity: FindingSeverity | null;
+  category: FindingCategory;
+  title: string;
+} {
+  const text = body.replace(/<!--\s*kodiai:review-output-key:[\s\S]*?-->/gi, "").trim();
+  const yamlMatch = text.match(/^```yaml\s*([\s\S]*?)```/i);
+
+  if (yamlMatch) {
+    const metadataLines = (yamlMatch[1] ?? "")
+      .split("\n")
+      .map((line) => line.trim())
+      .filter((line) => line.includes(":"));
+    const metadata = new Map<string, string>();
+    for (const line of metadataLines) {
+      const idx = line.indexOf(":");
+      if (idx <= 0) continue;
+      const key = line.slice(0, idx).trim().toLowerCase();
+      const value = line.slice(idx + 1).trim();
+      metadata.set(key, value);
+    }
+
+    const titleSection = text.slice(yamlMatch[0].length).trim();
+    const titleLine = titleSection
+      .split("\n")
+      .map((line) => line.trim())
+      .find((line) => line.length > 0) ?? "Untitled finding";
+    const title = titleLine.replace(/^\*\*(.+)\*\*$/, "$1").trim();
+
+    return {
+      severity: normalizeSeverity(metadata.get("severity")),
+      category: normalizeCategory(metadata.get("category")),
+      title,
+    };
+  }
+
+  const firstLine = text.split("\n").map((line) => line.trim()).find((line) => line.length > 0) ?? "";
+  const severityPrefix = firstLine.match(/^\[(critical|major|medium|minor)\]\s*(.*)$/i);
+  if (severityPrefix) {
+    return {
+      severity: normalizeSeverity(severityPrefix[1]),
+      category: "correctness",
+      title: (severityPrefix[2] || "Untitled finding").trim(),
+    };
+  }
+
+  return {
+    severity: null,
+    category: "correctness",
+    title: firstLine || "Untitled finding",
+  };
+}
