+++ src/lib/review-profile-presets.ts
@@ -1,0 +1,27 @@
+import type { FindingSeverity, ReviewArea } from "./review-finding-metadata.ts";
+
+export const PROFILE_PRESETS: Record<string, {
+  severityMinLevel: FindingSeverity;
+  maxComments: number;
+  ignoredAreas: ReviewArea[];
+  focusAreas: ReviewArea[];
+}> = {
+  strict: {
+    severityMinLevel: "minor",
+    maxComments: 15,
+    ignoredAreas: [],
+    focusAreas: [],
+  },
+  balanced: {
+    severityMinLevel: "medium",
+    maxComments: 7,
+    ignoredAreas: ["style"],
+    focusAreas: [],
+  },
+  minimal: {
+    severityMinLevel: "major",
+    maxComments: 3,
+    ignoredAreas: ["style", "documentation"],
+    focusAreas: ["security", "correctness"],
+  },
+};
