+++ src/lib/review-trigger-utils.ts
@@ -1,0 +1,19 @@
+export function isReviewTriggerEnabled(
+  action: string,
+  triggers: {
+    onOpened: boolean;
+    onReadyForReview: boolean;
+    onReviewRequested: boolean;
+    onSynchronize?: boolean;
+  },
+): boolean {
+  if (action === "opened") return triggers.onOpened;
+  if (action === "ready_for_review") return triggers.onReadyForReview;
+  if (action === "review_requested") return triggers.onReviewRequested;
+  if (action === "synchronize") return triggers.onSynchronize ?? false;
+  return false;
+}
+
+export function normalizeReviewerLogin(login: string): string {
+  return login.trim().toLowerCase().replace(/\[bot\]$/i, "");
+}
